+++ ricaun.RevitTest.Command/Extensions/EncodeParameterArgumentExtension.cs
@@ -1,0 +1,104 @@
+﻿using System;
+using System.Text;
+using System.Text.RegularExpressions;
+
+namespace ricaun.RevitTest.Command.Extensions
+{
+    /// <summary>
+    /// EncodeParameterArgumentExtension
+    /// </summary>
+    /// <remarks>
+    /// https://stackoverflow.com/questions/5510343/escape-command-line-arguments-in-c-sharp/6040946#6040946
+    /// </remarks>
+    internal static class EncodeParameterArgumentExtension
+    {
+        /// <summary>
+        ///     This routine appends the given argument to a command line such that
+        ///     CommandLineToArgvW will return the argument string unchanged. Arguments
+        ///     in a command line should be separated by spaces; this function does
+        ///     not add these spaces.
+        /// </summary>
+        /// <param name="argument">Supplies the argument to encode.</param>
+        /// <param name="force">
+        ///     Supplies an indication of whether we should quote the argument even if it 
+        ///     does not contain any characters that would ordinarily require quoting.
+        /// </param>
+        /// <remarks>https://learn.microsoft.com/en-us/archive/blogs/twistylittlepassagesallalike/everyone-quotes-command-line-arguments-the-wrong-way</remarks>
+        internal static string EncodeParameterArgumentW(this string argument, bool force = false)
+        {
+            if (argument == null) throw new ArgumentNullException(nameof(argument));
+
+            // Unless we're told otherwise, don't quote unless we actually
+            // need to do so --- hopefully avoid problems if programs won't
+            // parse quotes properly
+            if (force == false
+                && argument.Length > 0
+                && argument.IndexOfAny(" \t\n\v\"".ToCharArray()) == -1)
+            {
+                return argument;
+            }
+
+            var quoted = new StringBuilder();
+            quoted.Append('"');
+
+            var numberBackslashes = 0;
+
+            foreach (var chr in argument)
+            {
+                switch (chr)
+                {
+                    case '\\':
+                        numberBackslashes++;
+                        continue;
+                    case '"':
+                        // Escape all backslashes and the following
+                        // double quotation mark.
+                        quoted.Append('\\', numberBackslashes * 2 + 1);
+                        quoted.Append(chr);
+                        break;
+                    default:
+                        // Backslashes aren't special here.
+                        quoted.Append('\\', numberBackslashes);
+                        quoted.Append(chr);
+                        break;
+                }
+                numberBackslashes = 0;
+            }
+
+            // Escape all backslashes, but let the terminating
+            // double quotation mark we add below be interpreted
+            // as a metacharacter.
+            quoted.Append('\\', numberBackslashes * 2);
+            quoted.Append('"');
+
+            return quoted.ToString();
+        }
+
+        // This is an EDIT
+        // Note that this version does the same but handles new lines in the arugments
+        internal static string EncodeParameterArgumentMultiLine(this string original)
+        {
+            if (string.IsNullOrEmpty(original))
+                return original;
+            string value = Regex.Replace(original, @"(\\*)" + "\"", @"$1\$0");
+            value = Regex.Replace(value, @"^(.*\s.*?)(\\*)$", "\"$1$2$2\"", RegexOptions.Singleline);
+
+            return value;
+        }
+
+        /// <summary>
+        /// Encodes an argument for passing into a program
+        /// </summary>
+        /// <param name="original">The value that should be received by the program</param>
+        /// <returns>The value which needs to be passed to the program for the original value 
+        /// to come through</returns>
+        internal static string EncodeParameterArgument(this string original)
+        {
+            if (string.IsNullOrEmpty(original))
+                return original;
+            string value = Regex.Replace(original, @"(\\*)" + "\"", @"$1\$0");
+            value = Regex.Replace(value, @"^(.*\s.*?)(\\*)$", "\"$1$2$2\"");
+            return value;
+        }
+    }
+}